% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_detector.R
\name{apply_detector}
\alias{apply_detector}
\title{Use an event detector to generate labels}
\usage{
apply_detector(data, detector, ...)
}
\arguments{
\item{data}{a sumsarizer formatted data table for one or more sensor missions}

\item{detector}{a sumsarizer event detector function. see \link{event_detectors}}

\item{...}{arguments passed to the detector}
}
\description{
Use an event detector to generate labels
}
\seealso{
Other event_detectors: \code{\link{constant_detector}},
  \code{\link{firefinder_detector}},
  \code{\link{sl3_model_detector}},
  \code{\link{threshold_detector}}
}
\concept{event_detectors}
