% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{cooking_danny_algo}
\alias{cooking_danny_algo}
\title{Danny's Magic Cooking Algorithm
TODO: Document
TODO: use set instead of :=}
\usage{
cooking_danny_algo(one_file, sample_interval)
}
\description{
Danny's Magic Cooking Algorithm
TODO: Document
TODO: use set instead of :=
}
