% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_example_data.R
\name{download_example_data}
\alias{download_example_data}
\title{Download Example Sumsarizer Data Used in Vignette}
\usage{
download_example_data(path = "example_data")
}
\description{
Download Example Sumsarizer Data Used in Vignette
}
