% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{est_sample_interval}
\alias{est_sample_interval}
\title{Estimate Sample Interval
Estimates a sample interval in seconds as the median interval between adjacent samples}
\usage{
est_sample_interval(timestamp)
}
\arguments{
\item{timestamp}{a vector of timestamps}
}
\description{
Estimate Sample Interval
Estimates a sample interval in seconds as the median interval between adjacent samples
}
