% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{event_summaries}
\alias{event_summaries}
\title{Summarize Events}
\usage{
event_summaries(events)
}
\arguments{
\item{events}{a table of events as generated by \code{\link{list_events}}}
}
\description{
Generate basic event summaries
}
