% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_sample_interval}
\alias{get_sample_interval}
\title{Get Sample Interval, or estimate if not provided}
\usage{
get_sample_interval(data)
}
\arguments{
\item{data}{a sumsarizer formatted data table for one sensor mission}
}
\description{
Get Sample Interval, or estimate if not provided
}
