% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_auto_event_maker.R
\name{hapin_cooking_event_detector}
\alias{hapin_cooking_event_detector}
\title{HAPIN Cooking Event Detector}
\usage{
hapin_cooking_event_detector(data, threshold = 75, dgap = TRUE)
}
\arguments{
\item{data}{Dataframe or datatable containing at minimum timestamp and value columns}

\item{threshold}{Numeric temperature above which points are labeled as cooking}

\item{dgap}{Logical. When true, if the rolling difference in times exceeds the sampling interval in seconds, any points labeled as cooking as relabeled as non-cooking.}
}
\description{
An algorithm developed as part of the Household Air Pollution Intervention Network Randomized Control Trial.
}
