% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{import_folder}
\alias{import_folder}
\title{Import Folder of files and concatenate}
\usage{
import_folder(folder, ...)
}
\arguments{
\item{folder}{Path to folder}

\item{...}{extra arguments to sensor-specific read methods}
}
\description{
Import Folder of files and concatenate
}
