% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{import_sums}
\alias{import_sums}
\title{Generic SUMs import functions}
\usage{
import_sums(input_file, ...)
}
\arguments{
\item{input_file}{Path to file being imported}

\item{...}{extra arguments to sensor-specific read methods}
}
\value{
A data.frame with two columns - timestamp (POSIXct) and value (numeric)
}
\description{
Uses string parsing to guess which type of sensor
is being imported -- currently imports all manner of iButton data,
Wellzion SSN-61, Lascar, and EME Systems KSUMs thermocouples.
}
