% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sl3_ml.R
\name{learn_labels}
\alias{learn_labels}
\title{Learn a sl3 Machine Learning Model from Labels}
\usage{
learn_labels(data, sl3_learner = NULL)
}
\arguments{
\item{data}{a sumsarizer formatted data table for one or more sensor missions}

\item{sl3_learner}{A sl3 learner, or Lrnr_xgboost if no learner is specified}
}
\description{
Learn a sl3 Machine Learning Model from Labels
}
