% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{list_events}
\alias{list_events}
\title{List Events}
\usage{
list_events(data, event = NULL)
}
\arguments{
\item{labeled_data}{dataframe of labeled events, with, at minimum, timestamp, value, and event_num}
}
\description{
Generate a list of events
}
