% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\docType{data}
\name{make_features}
\alias{make_features}
\alias{FEATURE_NAMES}
\title{Make Features to Predict Cooking
TODO: DOCUMENT
TODO: use set instead of :=}
\format{An object of class \code{character} of length 24.}
\usage{
make_features(one_file)

FEATURE_NAMES
}
\description{
Make Features to Predict Cooking
TODO: DOCUMENT
TODO: use set instead of :=
}
\keyword{datasets}
