% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{number_events}
\alias{number_events}
\title{Number Events
Generate unique identifiers for each event, based on indicator of being in an event.
Nonevent periods are labelled NA}
\usage{
number_events(event_indicator)
}
\arguments{
\item{event_indicator}{A binary vector, where TRUE indicates an event}
}
\description{
Number Events
Generate unique identifiers for each event, based on indicator of being in an event.
Nonevent periods are labelled NA
}
