% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sums.R
\name{plot_sums}
\alias{plot_sums}
\title{Plot a SUMs File}
\usage{
plot_sums(data, events = NULL)
}
\arguments{
\item{data}{a sumsarizer formatted data table for one sensor mission}

\item{events}{a table of events as generated by \code{\link{list_events}}. 
If \code{NULL}, will be generated automatically from the \code{label} column of \code{data}}
}
\description{
Plot a SUMs File
}
