% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainset_utils.R
\name{raw_sums_to_trainset}
\alias{raw_sums_to_trainset}
\title{Convert a Folder of Raw SUMs Files to TRAINSET Compatible Files}
\usage{
raw_sums_to_trainset(raw_sums_path, trainset_path)
}
\arguments{
\item{raw_sums_path}{the path to a folder of raw sums files}

\item{trainset_path}{the path to a folder where the trainset compatible files will go}
}
\description{
Convert a Folder of Raw SUMs Files to TRAINSET Compatible Files
}
