% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.ibutton.R
\name{read.ibutton}
\alias{read.ibutton}
\title{Read iButton Thermocron files}
\usage{
read.ibutton(input_file, timezone = "UTC", lablr_output = T)
}
\arguments{
\item{input_file}{Path to iButton file}

\item{timezone}{Timezone. Defaults to UTC.}

\item{lablr_output}{Output a file for use with LABLR.}
}
\value{
A data.frame with two columns - timestamp (POSIXct) and value (numeric)
}
\description{
This function reads in data downloaded from Maxim iButton Thermocrons. It assumes
input files are CSVs exported from Maxim's OneWireViwer. It returns a two column data.frame: timestamp 
(POSIXct) and value (numeric). The default timezone is UTC, but can be otherwise specified. See 
\code{\link{timezones}} for more details.
}
\seealso{
\code{\link{timezones}}
}
