% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.lascartc.R
\name{read.lascar}
\alias{read.lascar}
\title{Read a file from Lascar thermocouples}
\usage{
read.lascar(input_file, timezone = "UTC", lablr_output = T)
}
\arguments{
\item{input_file}{Path to the Lascar Thermocouple file to be read in}

\item{timezone}{Timezone. Defaults to UTC.}

\item{lablr_output}{Output a file for use with LABLR.}
}
\value{
A data.frame with two columns - timestamp (POSIXct) and value (numeric)
}
\description{
This function reads in data downloaded from Lascar thermocouple dataloggers. 
It returns a two column data.frame: timestamp (POSIXct) and value (numeric).
The default timezone is UTC, but can be otherwise specified. See \code{\link{timezones}}
for more details.
}
\seealso{
\code{\link{timezones}}
}
