% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.wellzion.R
\name{read.wellzion}
\alias{read.wellzion}
\title{Read a Wellzion SSN61 file}
\usage{
read.wellzion(input_file, timezone = "UTC", lablr_output = T)
}
\arguments{
\item{input_file}{Path to the Wellzion SSN61 file to be read in}

\item{timezone}{Timezone. Defaults to UTC.}

\item{lablr_output}{Output a file for use with LABLR.}
}
\value{
A data.frame with two columns - timestamp (POSIXct) and value (numeric)
}
\description{
This function reads in data downloaded from a Wellzion SSN-61. It assumes
input files are CSVs exported from Wellzion's Data Logger Graph version 6.5
or greater. It returns a two column data.frame: timestamp (POSIXct) and value (numeric).
The default timezone is UTC, but can be otherwise specified. See \code{\link{timezones}}
for more details.
}
\seealso{
\code{\link{timezones}}
}
