% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ksums.R
\name{read_ksums}
\alias{read_ksums}
\title{Read EME Systems kSUMs}
\usage{
read_ksums(input_file, timezone = "UTC", trainset_output = F,
  wide = F)
}
\arguments{
\item{input_file}{Path to kSUMs file}

\item{timezone}{Timezone. Defaults to UTC.}

\item{trainset_output}{Output a file for use with TRAINSET with four columns: filename, timestamp, value, label}

\item{wide}{Specifies the format of the returned data -- either a wide data.frame or a list 
of data.frames (one for each sensing element).}
}
\value{
A data.frame with five columns - timestamp (POSIXct), unit (character), value (numeric), filename (character), and label.
}
\description{
This function reads in data downloaded from EME Systems kSUMS. 
It returns a five column data.frame: timestamp (POSIXct), unit (character), value (numeric), filename (character), 
and label. The default timezone is UTC, but can be otherwise specified. See \code{\link{timezones}} for more details.
}
\seealso{
\code{\link{timezones}}
}
