% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_detectors.R
\name{sl3_model_detector}
\alias{sl3_model_detector}
\title{Use sl3 Machine Learning for event detection}
\usage{
sl3_model_detector(data, model_obj = NULL, threshold = 0.5)
}
\arguments{
\item{data}{a sumsarizer formatted data table for one sensor mission}

\item{model_obj}{either a sl3 learner fit object, or a url or path to a .rdata file containing one}

\item{threshold}{a value between 0 and 1 indicating the sensitivity of the event detector}

\item{...}{not currently used}
}
\description{
Uses a sl3 machine learner model trained on labels from TRAINSET to detect events. 
See TODO to train your own
}
\seealso{
Other event_detectors: \code{\link{constant_detector}},
  \code{\link{firefinder_detector}},
  \code{\link{threshold_detector}}
}
\concept{event_detectors}
