% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sl3_ml.R
\name{sl3_task_from_data}
\alias{sl3_task_from_data}
\title{Construct a sl3 Task From SUMs Data}
\usage{
sl3_task_from_data(data)
}
\arguments{
\item{data}{a sumsarizer formatted data table for one or more sensor missions}
}
\description{
Construct a sl3 Task From SUMs Data
}
