% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{smooth_events}
\alias{smooth_events}
\title{Smooth Events}
\usage{
smooth_events(label, sample_interval, min_event_sec = 5 * 60,
  min_break_sec = 30 * 60)
}
\arguments{
\item{label}{A binary vector, where TRUE indicates an event}

\item{sample_interval}{the sample interval in seconds}

\item{min_event_sec}{minmum number of seconds in a real event}

\item{min_break_sec}{minmum number of seconds in a real non-event}
}
\description{
Reduce "blipiness" in event indicators by eliminating small cooking events and gaps
}
