% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{summarize_events}
\alias{summarize_events}
\title{Summarize Events}
\usage{
summarize_events(event_num, sample_interval_secs)
}
\arguments{
\item{event_num}{an event identifer as generated by \link{number_events}}

\item{sample_interval}{the sample interval in seconds}
}
\description{
Generate basic event summaries
}
