% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_detectors.R
\name{threshold_detector}
\alias{threshold_detector}
\title{Detect Cooking Using Simple Threshold}
\usage{
threshold_detector(data, threshold = 75, direction = ">", ...)
}
\arguments{
\item{data}{a sumsarizer formatted data table for one sensor mission}

\item{threshold}{the main threshold to determine cooking}

\item{direction}{one of > < >= or <=}

\item{...}{not currently used}
}
\description{
Detects cooking using a simple threshold.
}
\seealso{
Other event_detectors: \code{\link{apply_detector}},
  \code{\link{constant_detector}},
  \code{\link{firefinder_detector}},
  \code{\link{sl3_model_detector}}
}
\concept{event_detectors}
