% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{import_lablr}
\alias{import_lablr}
\title{Read labeled files from LABLR}
\usage{
import_lablr(input_file, metadata = NA, ...)
}
\arguments{
\item{input_file}{Path to file being imported}

\item{metadata}{Path to file containing metadata for imported files or NA}

\item{...}{not currently used}
}
\value{
A data.frame with at least four columns - timestamp (POSIXct), value (numeric), label (logical), filename (character)
}
\description{
Imports labeled files exported from LABLR
and merges with metadata if specified for further use by SUMSarizer
}
