% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\docType{data}
\name{sumsarizer_feature_names}
\alias{sumsarizer_feature_names}
\alias{make_features}
\title{Make Features to Predict Cooking
TODO: DOCUMENT}
\format{An object of class \code{character} of length 24.}
\usage{
sumsarizer_feature_names

make_features(timestamps, values, sample_interval = NULL)
}
\arguments{
\item{timestamps}{a vector of POSIXct timestamps}

\item{values}{a vector of cooking temperatures}

\item{sample_interval}{the sample interval in seconds}
}
\description{
Make Features to Predict Cooking
TODO: DOCUMENT
}
\keyword{datasets}
