% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ksums.R
\name{read_ksums}
\alias{read_ksums}
\title{Read EME Systems kSUMs}
\usage{
read_ksums(input_file, timezone = "UTC", lablr_output = T, wide = F)
}
\arguments{
\item{input_file}{Path to kSUMs file}

\item{timezone}{Timezone. Defaults to UTC.}

\item{lablr_output}{Output a unique file for each thermocouple lead for use with LABLR.}

\item{wide}{Specifies the format of the returned data -- either a wide data.frame or a list 
of data.frames (one for each sensing element).}
}
\value{
A data.frame with three columns - timestamp (POSIXct) and value (numeric); One for each thermocouple probe.
}
\description{
This function reads in data downloaded from EME Systems kSUMS. 
It returns a two column data.frame: timestamp (POSIXct) and value (numeric). 
The default timezone is UTC, but can be otherwise specified. See \code{\link{timezones}} for more details.
}
\seealso{
\code{\link{timezones}}
}
